// src/common/public-id/public-id-prefix.enum.ts
export enum PublicIdPrefix {
  USER = 'us',
  REQUEST = 'rq',
  TRAVEL = 'tr',
  DEMAND = 'dm',
  TRANSACTION = 'tx',
  REVIEW = 'rv',
  MESSAGE = 'ms',
  NOTIFICATION = 'nt',
  ALERT = 'at',
  BOOKMARK = 'bm',
  SUPPORT_REQUEST = 'sp',
  AIRPORT = 'ap',
  AIRLINE = 'al',
  CURRENCY = 'cu',
  UPLOADED_FILE = 'uf',
  DELIVERY_PROOF = 'dp',
}